accounts/app/requirements.txt
Django==3.0
django-cors-headers==3.7.0
djangorestframework==3.12.2
requests==2.25.1
